[add] quiz form base template
