+++ frontend/src/features/quiz-form/components/description/index.tsx
@@ -17,8 +17,7 @@ export const Description: React.FC = () => {
           <p className="text-gray-600 max-w-2xl text-sm">
             Readumは生成AIテクノロジーを活用したクイズ生成アプリケーションです。
             <br></br>
-            あなたの読書メモや学習内容をフォームに入力して送信するだけで、AIが自動的にあなたの理解度を測るための
-            オリジナルクイズを作成してくれます。
+            あなたの読書メモや学習内容をフォームに入力して送信するだけで、AIが自動的にあなたの理解度を測るためのオリジナルクイズを作成してくれます。
           </p>
         </div>
       </section>
